move overlapping model to own file
